--- app/src/main/java/com/techcourse/controller/RegisterViewController.java
@@ -1,13 +1,0 @@
-package com.techcourse.controller;
-
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-public class RegisterViewController implements Controller {
-
-    @Override
-    public String execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
-        return "/register.jsp";
-    }
-}
